feat: 判断是否在待回收模块逻辑调整 (merge request !2078)
Squash merge branch 'v1.7.x-adjust-recycle-module' into 'bcc/v1.7.x'
feat: 判断是否在待回收模块逻辑调整

 



TAPD: --story=122275137

+++ cmd/woa-server/logics/task/recycler/returner/returner.go
@@ -552,7 +552,7 @@ func (r *Returner) getBizRecycleModuleID(bizID int64) (int64, error) {
 
 	moduleID := int64(0)
 	for _, module := range resp.Data.Module {
-		if module.Default == int64(cmdb.DftModuleRecycle) {
+		if module.Default == cmdb.DftModuleRecycle {
 			moduleID = module.BkModuleId
 			break
 		}
